requirements.txt
click==7.0
confuse
readsettings==3.4.0
web3==4.9.2
texttable==1.6.2
python-dateutil==2.8.0
